+++ backend/src/users/users.controller.ts
@@ -1,12 +1,7 @@
-import { existsSync, mkdirSync } from 'fs';
-import { extname } from 'path';
 import {
   Controller,
-  FileTypeValidator,
   Get,
-  MaxFileSizeValidator,
   Param,
-  ParseFilePipe,
   ParseUUIDPipe,
   Post,
   StreamableFile,
@@ -15,10 +10,8 @@ import {
   UseInterceptors,
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
-import { MulterOptions } from '@nestjs/platform-express/multer/interfaces/multer-options.interface';
 import { User } from '@prisma/client';
-import { Request } from 'express';
-import { diskStorage } from 'multer';
+import { FileService } from 'src/file/file.service';
 import { GetUser } from '../auth/decorator/get-user.decorator';
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 import { PostGuard } from './guards/post.guard';
@@ -26,7 +19,10 @@ import { UsersService } from './users.service';
 
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(
+    private readonly usersService: UsersService,
+    private readonly fileService: FileService
+  ) {}
 
   @Get('all')
   @UseGuards(JwtAuthGuard)
@@ -40,37 +36,14 @@ export class UsersController {
     return user;
   }
 
-  static readonly multerOptions = (): MulterOptions => ({
-    storage: diskStorage({
-      destination: (req: Request & { user?: { user?: User } }, file, cb) => {
-        const dir = `./upload/${req.user?.user?.id ?? ''}/`;
-        if (!existsSync(dir)) {
-          mkdirSync(dir, { recursive: true });
-        }
-        cb(null, dir);
-      },
-      filename: (req: Request & { user?: { user?: User } }, file, cb) => {
-        cb(null, `${req.user?.user?.name ?? ''}${extname(file.originalname)}`);
-      },
-    }),
-  });
-
-  static readonly parseFilePipe = (): ParseFilePipe =>
-    new ParseFilePipe({
-      validators: [
-        new MaxFileSizeValidator({ maxSize: 10000000 }),
-        new FileTypeValidator({ fileType: /jpeg|png|jpg/ }),
-      ],
-    });
-
   @Post(':id/avatar')
   @UseGuards(JwtAuthGuard, PostGuard)
-  @UseInterceptors(FileInterceptor('file', UsersController.multerOptions()))
+  @UseInterceptors(FileInterceptor('file', FileService.multerOptions()))
   async uploadFile(
     @GetUser() user: User,
-    @UploadedFile(UsersController.parseFilePipe()) file: Express.Multer.File
+    @UploadedFile(FileService.parseFilePipe()) file: Express.Multer.File
   ): Promise<User> {
-    this.usersService.deleteOldFile(file.filename, user);
+    this.fileService.deleteOldFile(file.filename, user);
 
     const updateColumns = {
       avatarUrl: `http://localhost:3000/users/${user.id}/avatar/${file.filename}`,
@@ -87,6 +60,6 @@ export class UsersController {
   ): StreamableFile {
     const path = `./upload/${id}/${filename}`;
 
-    return this.usersService.streamAvatar(path);
+    return this.fileService.streamFile(path);
   }
 }
