+++ backend/src/users/users.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
+import { FileModule } from 'src/file/file.module';
 import { AuthModule } from '../auth/auth.module';
 import { PrismaModule } from '../prisma/prisma.module';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 
 @Module({
-  imports: [AuthModule, PrismaModule],
+  imports: [AuthModule, PrismaModule, FileModule],
   controllers: [UsersController],
   providers: [UsersService],
 })
